movestack
This plugin allows you to move clients around in the stack and swap them
with the master.  It emulates the behavior off mod+shift+j and
mod+shift+k in Xmonad.

movestack(+1) will swap the client with the current focus with the next
client.

movestack(-1) will swap the client with the current focus with the
previous client.

URL: http://dwm.suckless.org/patches/movestack/dwm-movestack-6.1.diff