movestack
This plugin allows you to move clients around in the stack and swap them
with the master.  It emulates the behavior off ``MODKEY+Shift+j`` and
``MODKEY+Shift+k`` in Xmonad_.

``movestack(+1)`` will swap the client with the current focus with the
next client.

``movestack(-1)`` will swap the client with the current focus with the
previous client.

.. _Xmonad: https://xmonad.org

URL: http://dwm.suckless.org/patches/movestack/dwm-movestack-6.1.diff