rotatestack
Stack rotation moves a client from the bottom to the top of the stack
(or the other way round). This effectively rotates the clients by one
position clockwise (or CCW, respectively).

It should play well with arbitrary stack layouts and nmaster values.

One may think of it as moving the zoom through the list of clients, very
much in the same way as scrolling moves the view port around a pane.

URL: http://dwm.suckless.org/patches/rotatestack/dwm-rotatestack-20161021-ab9571b.diff